+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Canonical reference for changes, improvements, and bugfixes for cap.
 
 ## Next
+
+* feat (jwt): add optional callback to dynamically fetch key sets in validator ([PR #187](https://github.com/hashicorp/cap/pull/187))
 * fix (saml): always validate response and assertion signatures when signatures are present ([PR #180](https://github.com/hashicorp/cap/pull/180))
 
 ## 0.12.0